awesome_libs v0.0.1 AwesomeLibs.CLI

Handle the command line parsing and dispatch to the correct functions.

Return a tuple of {mode, filter} where mode can :category or :library or :help if help was given

Functions

parse_args(argv)

process(arg1)

Processes a category search request and returns the filtered category list.

##Example

iex> AwesomeLibs.CLI.process {:category, “2”} :ok

run(argv)