awesome_libs v0.0.1 AwesomeLibs.MdParser

 Basic MarkDown parser which only parses the list parts of the document, and converts it to a strucut list.

Functions

load()

 It loads the stored JSON structure and returns the deserialize it. Warn: the process method should be called before

process(url)

Fetch the given url, parse the markdown file and saves the result in JSON format.

Example

iex> AwesomeLibs.MdParser.process "https://raw.githubusercontent.com/h4cc/awesome-elixir/master/README.md"
  :ok

reformat(lines)

Reformats the given list of strings and converts them to LibItem. It expects thaht all the strings are markdown lists, so they are in a format:

some text